% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeModel.R
\name{AgeModel}
\alias{AgeModel}
\title{Plots the Pb-210 age model and validation information}
\usage{
AgeModel(
  List = List,
  PlotName = "Age_Model",
  DepthMin = NA,
  DepthMax = NA,
  TimeMin = 1900,
  TimeMax = NA,
  SavePNG = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{List}{Dating object, created by \code{\link{ReadData}} and transformed 
with the package functions.}

\item{PlotName}{Name of the plot (without extension) to be produced. 
The extension will automatically be set to \code{.png}. Default is \code{"Age Model"}.}

\item{DepthMin}{Minimum depth (cm) of the age model plot. Default is \code{NA}.}

\item{DepthMax}{Maximum depth (cm) of the age model plot. Default is \code{NA}.}

\item{TimeMin}{Minimum time (decimal age) of the age model plot. Default is \code{NA}.}

\item{TimeMax}{Maximum time (decimal age) of the age model plot. Default is \code{NA}.}

\item{SavePNG}{Logical. If TRUE, saves a high quality figure.}

\item{verbose}{Logical. If TRUE, prints progress logs to the console.}

\item{...}{Extra arguments passed to plotting functions (e.g., \code{\link[graphics]{par}}).}
}
\value{
Generates a \code{.PNG} plot of the age model with validation ellipses
and p-values. 
The plot is also reproduced in the R graphics device (e.g., RStudio plots pane).
}
\description{
This function plots the age model after \eqn{^{210}}Pb dating and, if provided, 
the validation information (e.g., independent chronological markers). 
It also provides the statistical significance of the validation.
}
\details{
The age model is a representation of age (year CE) versus depth. 
For each point, both uncertainty bars are plotted. 
Minimum and maximum values for plotting are automatically calculated if not 
explicitly provided. 

For validation, two approaches are used. 
For visual validation, 1-\eqn{\sigma} (68\%, red) and 2-\eqn{\sigma} (95\%, pink) 
ellipses of each validation point are plotted. 
The validation ellipses are calculated assuming that the 
variables (year and depth of the event) follow a uniform (square) distribution. 

The validation ellipses include the user-provided label. 
As in most cases these will correspond to radionuclides, numbers are superscript,
so we recommend using the format "123Element" (e.g. 137Cs). 

For statistical validation, a significance (p-value) of the Mahalanobis distance between the validation point and each age-model curve, is also provided.
}
\examples{
\dontrun{
a <- ReadData()
a <- Equilibrium(a, interactive = FALSE)
a.cf <- Pb210CF(a)
AgeModel(a.cf)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{Pb210CF}}, \code{\link{Pb210CFCS}}
}
