% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfcs.R
\name{cfcs}
\alias{cfcs}
\title{Empty dating function}
\usage{
cfcs(total_col, support_col, sample_date = NULL, constants = NULL,
  n_sect = length(s_mass), constant = NULL, runs = 1e+05)
}
\arguments{
\item{constant}{A NULL value when measured supported values are used to calculate excess 210Pb.  If a constant background value is used, constant must be a two element numeric vector with the value and uncertainty.}

\item{runs}{The number of simulations to run for the MCMC.}

\item{inv_top}{In cases where a user may believe that some of the upper core is missing, approximate the missing proportion inventory.}

\item{inv_bot}{In cases where 210Pb does not meet the support values, approximate the missing proportion inventory to reach equilibrium.}
}
\description{
Add stuff, but include that all uncertinty should be reported as one sigma, or instrument resolution.
}
