% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompleteProfile.R
\name{CompleteProfile}
\alias{CompleteProfile}
\title{Complete missing data of sediment core profiles for the CF model}
\usage{
CompleteProfile(List, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{verbose}{Logical.Print progress messages. Default is \code{TRUE}.}
}
\value{
Updated dating list with complete profiles.
}
\description{
Completes missing values in sediment core profiles used for \eqn{^{210}}Pb radiochronology.
Only interpolation (not extrapolation) is performed. It leaves original data intact.
}
\details{
Required top values: \code{MassSection} and \code{Pb210ExcessDecay}.
\code{Ra226} is optional. If no \eqn{^{226}}Ra data is provided, it is skipped with a warning.

Interpolation rules:
\itemize{
  \item \code{MassSection}: Linear interpolation vs. mean depth.
  \item \code{Ra226}: Linear interpolation vs. cumulative mass.
  \item \code{Pb210Total}: Exponential interpolation vs. cumulative mass.
  \item \code{Pb210ExcessDecay}: Exponential interpolation vs. cumulative mass.
}
}
\examples{
\dontrun{
List <- ReadData()
List <- CompleteProfile(List)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{Pb210CF}}.
}
