% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantRa.R
\name{ConstantRa}
\alias{ConstantRa}
\title{Calculates a constant supported Ra-226 activity}
\usage{
ConstantRa(List, mode = "auto", verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{mode}{Character. Total activity used to determine the constant Ra-226 activity.: 
\code{"auto"} (default), \code{"Pb"} or \code{"Po"}. 
If both \eqn{^{210}}Pb and \eqn{^{210}}Po are present, the user \strong{must} specify which one to use.}

\item{verbose}{Logical. Shows estimation details (default \code{TRUE}).}
}
\value{
Updated dating list with constant \eqn{^{226}}Ra and its uncertainty.
}
\description{
If no \eqn{^{226}}Ra values are available, this function estimates a constant 
supported (base) activity from the bottom-most \eqn{^{210}}Po or \eqn{^{210}}Pb total activities.
}
\details{
At least three non-NA values are required.

The algorithm starts from the core bottom and, if needed, progressively includes 
more sections to calculate the mean and standard deviation. It stops when 
the activity of the section immediately above is significantly higher 
than the current mean + 2 SD (indicating the equilibrium zone has been found).

The resulting mean is assigned as constant \code{Ra226} for the entire profile.

\strong{Important:} The user should carefully review the result. 
This function only handles simple cases. If the profile does not reach 
equilibrium, the last value is used as fallback but it is likely wrong.
}
\examples{
\dontrun{
  List <- ReadData()
  List <- ConstantRa(List, mode = "Po")}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{ReadData}}, \code{\link{DecayCorrection}}, 
\code{\link{Equilibrium}}, \code{\link{EquilibriumSet}}
}
