% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Load constants for dating evaluation.}
\usage{
constants(file, raw_data, total_col, support_col, NSectionDating = NULL)
}
\arguments{
\item{file}{Local file from which to load constants.}
}
\description{
Load in datasets and associated data to help construct age models.
}
