% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equilibrium.R
\name{Equilibrium}
\alias{Equilibrium}
\title{Calculate the bottom layer with no excess Pb-210 below}
\usage{
Equilibrium(List, Interactive = TRUE, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{Interactive}{Logical. If \code{TRUE} (default), allows interactive visual selection 
of the equilibrium layer (highly recommended). If \code{FALSE}, it uses the automatic rule.}

\item{verbose}{Logical. If TRUE, prints progress logs to the console (default \code{TRUE}).}
}
\value{
Updated dating list with the value of \code{Val$LayerDating}.
}
\description{
Estimates the layer below which there is no measurable excess \eqn{^{210}}Pb 
(the "equilibrium layer"). This value is critical for the CF dating model.

The function works after \code{\link{DecayCorrection}}, 
so the presence of \code{Pb210ExcessDecay} is expected.
}
\details{
Excess \eqn{^{210}}Pb is shown with 1-sigma uncertainty.
In interactive mode the x-axis shows tick marks and labels exactly centered 
under each red-blue bar pair. You can zoom in/out.
}
\examples{
\dontrun{
List <- Equilibrium(List)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{ConstantRa}}, \code{\link{EquilibriumSet}}, \code{\link{MissingInventory}}, 
\code{\link{Pb210CF}}, \code{\link{Pb210CFCS}}
}
