% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquilibriumSet.R
\name{EquilibriumSet}
\alias{EquilibriumSet}
\title{Manually set the equilibrium layer for Pb-210 dating}
\usage{
EquilibriumSet(List, LayerDating, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{\link{ReadData}}.}

\item{LayerDating}{Numeric. Bottom depth (cm) of the equilibrium layer. 
It \strong{must} be one of the values present in \code{Val$DepthLayerBottom}.}

\item{verbose}{Logical. If TRUE, prints progress logs to the console (default \code{TRUE}).}
}
\value{
Updated dating list with the value of \code{LayerDating} stored in 
  \code{$Val$LayerDating}.
}
\description{
Manually assigns the deepest layer below which there is no measurable excess \eqn{^{210}}Pb
(the "equilibrium layer"). This value is required for the dating models. 

Use this function when you prefer to override the automatic proposal or the interactive 
selection performed by \code{\link{Equilibrium}}.
}
\details{
The function performs a strict validation: \code{LayerDating} must match exactly 
one of the section bottom depths. No interpolation or nearest-neighbour matching is done.
}
\examples{
\dontrun{
List <- EquilibriumSet(List, LayerDating = "your_dating_layer_(cm)")}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{ConstantRa}}, \code{\link{Equilibrium}}, \code{\link{MissingInventory}}, 
\code{\link{Pb210CF}}, \code{\link{Pb210CFCS}}
}
