% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_interp.R
\name{exp_interp}
\alias{exp_interp}
\title{Fill missing data with exponential fit.}
\usage{
exp_interp(s_mass, po210, po201unc, pb210, pb210unc)
}
\arguments{
\item{s_mass}{The section mass for the sample, in order by depth from highest sample to deepest.}

\item{x}{A \section{vector} or \code{data.frame} to interpolate.}
}
