% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{Output}
\alias{Output}
\title{Output merged Pb-210 dating results (CF + CFCS) in a single CSV file}
\usage{
Output(List, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by
\code{Pb210CF} and/or \code{Pb210CFCS}.}

\item{verbose}{Logical.Print progress messages. Default is \code{TRUE}.}
}
\value{
Invisibly returns the final data.frame. Side effect: writes
  \code{"<CoreName> results.csv"} in the working directory.
}
\description{
Creates one single CSV file with per-section data in
and core-wide constants in the last four columns (`Name`, `Value`,
`Uncertainty`, `p-value`).
}
\examples{
\dontrun{
Output(List = List)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{ReadData}}, \code{\link{Pb210CF}}, 
\code{\link{Pb210CFCS}}, \code{\link{MissingInventory}}.
}
