% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pb210CF.R
\name{Pb210CF}
\alias{Pb210CF}
\title{Pb-210 dating of sediment cores with the Constant Flux (CF) model}
\usage{
Pb210CF(List, MonteCarloRunsCF = 2e+05, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{MonteCarloRunsCF}{Number of Monte Carlo iterations for uncertainty estimation.
Default is 20000. Use \code{MonteCarloRunsCF = 0} to skip uncertainty calculation.}

\item{verbose}{Logical.Print progress messages. Default is \code{TRUE}.}
}
\value{
Updated dating list. Results in $Val, uncertainties in $Unc.
}
\description{
Performs \eqn{^{210}}Pb dating using the Constant Flux (CF) model.
Based on Sanchez-Cabeza and Ruiz-Fernandez (2012).
}
\details{
Missing inventories at the top (\code{MissingInventoryTop}) and bottom
(\code{MissingInventoryBottom}) of the core are supported but their use requires
careful justification. The core is assumed to be cylindrical.
Uncertainties are propagated via Monte Carlo simulation
(Sanchez-Cabeza et al., 2014).
}
\examples{
\dontrun{
  List <- ReadData()
  List <- Pb210CF(List)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{CompleteProfile}}, \code{\link{Pb210CFCS}}, 
\code{\link{AgeModel}}, \code{\link{Equilibrium}}, \code{\link{EquilibriumSet}}, 
\code{\link{MissingInventory}}, \code{\link{Output}}, \code{\link{PlotMARSAR}}.
}
