% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pb210CFCS.R
\name{Pb210CFCS}
\alias{Pb210CFCS}
\title{Pb-210 dating of sediment cores with the CFCS model}
\usage{
Pb210CFCS(List, Breaks = NA, MonteCarloRunsCFCS = 10000, verbose = TRUE)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{Breaks}{In order to be as clear as possible about the user intention,
the vector must contain the following layers (section bottom depth):
\itemize{
\item{The surface layer (0).}
\item{The following intermediate break points (if any).}
\item{The depth until which fitting will take place.
Its maximum value is the layer down to which excess \eqn{^{210}}Pb is observed. 
It  can be last break point.}
The minimum number of elements of Breaks is 2, corresponding to the surface layer and
the end of the segment to be used for CFCS fitting.}}

\item{MonteCarloRunsCFCS}{Number of Monte Carlo iterations for uncertainty. Default = 10000.}

\item{verbose}{Logical. If TRUE, prints progress logs to the console (default \code{TRUE}).}
}
\value{
Updated dating list with CFCS results in $Val and uncertainties in $Unc.
}
\description{
Performs \eqn{^{210}}Pb dating using the Constant Flux and Constant Sedimentation (CFCS) model,
with optional piece-wise segments, although this option is not yet fully tested. 
The user should consider if the profile can meet simultaneoulsy 
the hypothesis of constant excess \eqn{^{210}}Pb flux and constant sedimentation.
In principal, the user must indicate to which depth the model can be used.
Although debatable, some authors use a piece-wise model (by segments),
and the user can also provide the proposed break points for the model.
However, one must take into account that, in this case,
the constant flux hypothesis is only maintained in each profile segment. 

When selecting the break sections, the user should be aware of the limitations of
linear regression analysis. This very much depends on the dataset and the expected robustness,
but we think that a linear regression of less than 5 points may usually be not statitiscally significant.

Please note the required contents of the Breaks() vector,
which also defines the maximum layer to which CFCS fitting will be done.
}
\examples{
\dontrun{
List <- Pb210CFCS(List, Breaks = c(0, "your_core_ dating_layer"))}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{Pb210CF}}, \code{\link{AgeModel}}, 
\code{\link{MissingInventory}}, \code{\link{Output}}.
}
