% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pb210Dating.R
\name{Pb210Dating}
\alias{Pb210Dating}
\title{Complete Pb-210 dating workflow}
\usage{
Pb210Dating(FileName = NA, LayerDating = NA, verbose = TRUE)
}
\arguments{
\item{FileName}{Character. CSV file containing the sediment-core data.
Default = NA uses the package example.}

\item{LayerDating}{Numeric. Bottom depth (cm) of the dating interval.
If NA (default), the equilibrium layer is selected automatically
using Equilibrium().}

\item{verbose}{Logical.Print progress messages. Default is \code{TRUE}.}
}
\value{
Dating list containing all calculated variables from the CF and CFCS
models.
}
\description{
Function wrapper to run the complete workflow for a sediment core suitable for \eqn{^{210}}Pb dating. 
This function is suitable for calculations with a well-known core, 
in which the equilibrium depth is well constrained. 
If run without careful core exploration, 
the risk of bad results is high and it is discouraged.
}
\examples{
\dontrun{
Pb210Dating(FileName = List, LayerDating = 17)}

}
