% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMARSAR.R
\name{PlotMARSAR}
\alias{PlotMARSAR}
\title{Plots accumulation rates of sediment cores dated with the CF model}
\usage{
PlotMARSAR(
  List = List,
  YearMin = 1900,
  YearMax = NA,
  MARMin = 0,
  MARMax = NA,
  SARMin = 0,
  SARMax = NA,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{List}{Dating object, created by \code{\link{ReadData}} and dated with \code{\link{Pb210CF}}.}

\item{YearMin}{Minimum year to be plotted. Default is \code{1900}.}

\item{YearMax}{Maximum year to be plotted (auto-calculated if NA).}

\item{MARMin}{Minimum MAR to be plotted. Default is \code{0}.}

\item{MARMax}{Maximum MAR to be plotted (auto-calculated if NA).}

\item{SARMin}{Minimum SAR to be plotted. Default is \code{0}.}

\item{SARMax}{Maximum SAR to be plotted (auto-calculated if NA).}

\item{verbose}{Logical. If TRUE, prints progress logs (default TRUE).}

\item{...}{Extra arguments passed to plotting functions.}
}
\value{
Two PNG files (`Core MAR.png` and `Core SAR.png`) + combined plot in the RStudio pane.
}
\description{
This function produces two independent PNG plots for the \strong{MassAccumulationRate} (MAR) 
and the \strong{SedimentAccumulationRate} (SAR) of cores previously dated with 
the Constant Flux (CF) model only. 
It uses the per-section rates calculated by \code{\link{Pb210CF}} (not the CFCS means).
}
\examples{
\dontrun{
After \code{\link{Pb210CF}} has been run: 
PlotMARSAR(List = List)}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
\eqn{^{210}}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{Pb210CF}}.
}
