% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProfiles.R
\name{PlotProfiles}
\alias{PlotProfiles}
\title{Plot of sediment profiles: flexible selection of Pb-210, Po-210, Ra-226 and proxy}
\usage{
PlotProfiles(
  List,
  DepthMin = NA,
  DepthMax = NA,
  PbMax = NA,
  RaMax = NA,
  ProxyMax = NA,
  Proxy = FALSE,
  NamePlot = "Profiles",
  SavePNG = TRUE,
  PlotActivities = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{List}{Dating list created by \code{ReadData} and processed by other package functions.}

\item{DepthMin}{Minimum depth to plot (cm). If NA, estimated automatically.}

\item{DepthMax}{Maximum depth to plot (cm). If NA, estimated automatically.}

\item{PbMax}{Axis limits when \code{PlotActivities = NULL}). 
If not NULL, axis limit is computed automatically from the selected radionuclides.}

\item{RaMax}{Same as above.}

\item{ProxyMax}{Same as above.}

\item{Proxy}{Logical. Needed when \code{PlotActivities = NULL}).}

\item{NamePlot}{Base name for the PNG file (default "Profiles").}

\item{SavePNG}{Logical. If TRUE (default), saves a high-resolution PNG file.}

\item{PlotActivities}{Character vector specifying which activities to plot on the 
main (left/top) axis. Possible values:  
\code{"Pb210Total"}, \code{"Pb210Excess"}, 
\code{"Pb210ExcessDecay}, \code{"Pb210ExcessDecayComplete} (this is the one used for dating), 
\code{"Po210Total"}, \code{"Po210Excess}, \code{"Po210ExcessDecay}, 
\code{"Po210ExcessDecayComplete}, 
\code{"Ra226}, \code{"Ra226Complete}, \code{"Proxy"}. 
The argument is \strong{case-insensitive} (e.g. "pb210excess", "PB210EXCESS", or "Pb210Excess" all work). 
Default \code{NULL} automatically identifies and plots 
\code{"Pb210Total"}, \code{"Po210Total"} and \code{"Ra226}.}

\item{verbose}{Logical. If TRUE, prints progress logs to the console (default \code{TRUE}).}

\item{...}{Further arguments passed to graphical parameters.}
}
\value{
Invisibly returns NULL. Side effects: optionally creates a .png file and displays a plot.
}
\description{
Plots sediment core profiles of any combination of total/excess/complete \eqn{^{210}Pb}, 
total/excess/complete \eqn{^{210}Po}, total/complete \eqn{^{226}Ra} 
and one proxy (e.g. \eqn{^{137}Cs}). 
The user can select which variables to plot via the \code{PlotActivities} 
argument. If left \code{NULL}, the function checks the input data 
and plots total values.
}
\examples{
\dontrun{
List <- ReadData() # Reads demonstration core data (TehuaII.rda)
PlotProfiles(List = List, SavePNG = TRUE, PlotActivities = "Po210Total")}

}
\references{
Sanchez-Cabeza, J. A. & Ruiz-Fernandez, A. C. (2012). 
^{210}Pb sediment radiochronology: an integrated formulation and 
classification of dating models. 
\emph{Geochimica et Cosmochimica Acta}, 82, 183-200.
}
\seealso{
\code{\link{ReadData}}, 
\code{\link{CompleteProfile}}, \code{\link{ConstantRa}}, 
\code{\link{DecayCorrection}}
}
