% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare4Dating.R
\name{Prepare4Dating}
\alias{Prepare4Dating}
\title{Prepare sediment-core data for Pb-210 dating}
\usage{
Prepare4Dating(NameFile = NA, verbose = TRUE)
}
\arguments{
\item{NameFile}{Character.
Name of the CSV file containing the sediment-core data.
If omitted or \code{NA}, the example dataset included in the
package is used.}

\item{verbose}{Logical.Print progress messages. Default is \code{TRUE}.}
}
\value{
A list containing original measurements, decay-corrected
radionuclide activities, excess \eqn{^{210}Pb} activities,
and completed excess \eqn{^{210}Pb} profiles.
}
\description{
Performs the standard preprocessing workflow required before
applying the \eqn{^{210}Pb} dating models.

This function is a convenience wrapper that sequentially executes:

\enumerate{
  \item \code{\link{ReadData}}
  \item \code{\link{DecayCorrection}}
  \item \code{\link{CompleteProfile}}
}

The objective is to generate the decay-corrected excess
\eqn{^{210}Pb} profiles required by the dating models.

Depending on the radionuclide information supplied by the user:

\itemize{
  \item If total \eqn{^{210}Pb} activities are provided,
        excess \eqn{^{210}Pb} activities are calculated,
        decay-corrected to the sampling date, and stored as
        \code{Pb210ExcessDecay}.

  \item If total \eqn{^{210}Po} activities are provided,
        activities are first corrected for the elapsed time
        between plating and measurement using the
        \eqn{^{210}Po} decay constant. The resulting activities
        are then decay-corrected to the sampling date using the
        \eqn{^{210}Pb} decay constant, assuming secular
        equilibrium between \eqn{^{210}Po} and
        \eqn{^{210}Pb}. Excess activities are stored as
        \code{Pb210ExcessDecay}.

  \item If (\eqn{^{226}Ra}) is not available,
        it is estimated using \code{\link{ConstantRa}} whenever possible.
}

Missing sections are subsequently interpolated against
cumulative dry mass depth using \code{\link{CompleteProfile}},
generating \code{Pb210ExcessDecayComplete}.

The resulting object contains:

\itemize{
  \item \code{Pb210ExcessDecay}, intended for the
        Constant Flux-Constant Sedimentation (CFCS) model.

  \item \code{Pb210ExcessDecayComplete}, intended for the
        Constant Flux (CF) model.
}

No dating model is executed.
The user must subsequently define the dating interval through
\code{\link{Equilibrium}} or \code{\link{EquilibriumSet}},
which are use to determine \code{LayerDating}.
}
\examples{
\dontrun{
# Standard preprocessing
List <- Prepare4Dating()
# Select dating interval
List <- Equilibrium(List)
# Run dating models
CFCS <- Pb210CFCS(List)
CF   <- Pb210CF(List)}

}
\references{
Sanchez-Cabeza, J.A. and Ruiz-Fernandez, A.C. (2012).
\emph{210Pb sediment radiochronology: an integrated formulation
and classification of dating models}.
Geochimica et Cosmochimica Acta, 82, 183--200.
}
\seealso{
\code{\link{ReadData}}, \code{\link{DecayCorrection}}, \code{\link{CompleteProfile}},
\code{\link{Equilibrium}}, \code{\link{EquilibriumSet}}, 
\code{\link{Pb210CFCS}}, \code{\link{Pb210CF}}
}
